Move UntypedFunctionUsageInspectionTests for QuickFixes into relevant files under RD.QuickFixes
Additionally ignore the tests moved out and clean up the
"RemoveOptionBaseDeclaration" strangeness in
"RemoveStopKeywordQuickFixTests"
