+++ RubberduckTests/QuickFixes/RemoveStopKeywordQuickFixTests.cs
@@ -34,9 +34,8 @@ End Sub";
             var inspection = new StopKeywordInspection(state);
             var inspector = InspectionsHelper.GetInspector(inspection);
             var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            // FIXME wat?
-            new RemoveOptionBaseStatementQuickFix(state).Fix(inspectionResults.First());
+            
+            new RemoveStopKeywordQuickFix(state).Fix(inspectionResults.First());
             Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
